+++ pkg/server/deploy/responses.go
@@ -1,6 +1,8 @@
 package deploy
 
 import (
+	"strconv"
+
 	dpb "github.com/luizalabs/teresa/pkg/protobuf/deploy"
 )
 
@@ -9,6 +11,30 @@ type ReplicaSetListItem struct {
 	Description string
 	Age         int64
 	Current     bool
+}
+
+type ByRevision []*dpb.ListResponse_Deploy
+
+func (s ByRevision) Len() int {
+	return len(s)
+}
+
+func (s ByRevision) Swap(i, j int) {
+	s[i], s[j] = s[j], s[i]
+}
+
+func (s ByRevision) Less(i, j int) bool {
+	k, err := strconv.Atoi(s[i].Revision)
+	if err != nil {
+		return false
+	}
+
+	l, err := strconv.Atoi(s[j].Revision)
+	if err != nil {
+		return false
+	}
+
+	return k < l
 }
 
 func newListResponse(items []*ReplicaSetListItem) *dpb.ListResponse {