+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Next Version] - Release Date
+## [0.8.0] - 2017-09-18
 ### Added
 - Command 'deploy list'
 - Command 'team remove-user'